[Restart] model설정부터 다시
