--- SWIDEA_SITE/IdeaBlog/forms.py
@@ -1,9 +1,0 @@
-from django import forms
-from .models import Idea,
-
-
-class PostForm(forms.ModelForm):
-
-    class Meta:
-        model = Idea
-        fields = '__all__'
