--- SWIDEA_SITE/IdeaBlog/views.py
@@ -1,23 +1,0 @@
-from django.shortcuts import render
-from .models import Idea
-# Create your views here.
-
-
-def idea_list(request):
-    pass
-
-
-def idea_detail(request, pk):
-    pass
-
-
-def create_idea(request):
-    pass
-
-
-def edit_idea(request):
-    pass
-
-
-def delete_idea(request, pk):
-    pass
