+++ SWIDEA_SITE/config/settings.py
@@ -38,8 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'devtoolBlog',
-    'ideaBlog'
+    'devTools',
+    'ideas',
 ]
 
 MIDDLEWARE = [
