+++ SWIDEA_SITE/devTools/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class DevtoolblogConfig(AppConfig):
+class DevtoolsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'devtoolBlog'
+    name = 'devTools'
