+++ SWIDEA_SITE/devTools/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path, include
+
+urlpatterns = [
+
+]
