--- SWIDEA_SITE/devtoolBlog/forms.py
@@ -1,8 +1,0 @@
-from django import forms
-from .models import Devtool
-
-
-class ToolForm(forms.ModelForm):
-    class Meta:
-        model = Devtool
-        fields = '__all__'
