--- SWIDEA_SITE/devtoolBlog/models.py
@@ -1,12 +1,0 @@
-from django.db import models
-
-# Create your models here.
-
-
-class Devtool(models.Model):
-    name = models.CharField(max_length=20)
-    kind = models.CharField(max_length=50)
-    description = models.TextField()
-
-    def __str__(self):
-        return self.name
