--- SWIDEA_SITE/devtoolBlog/templates/detail.html
@@ -1,16 +1,0 @@
-DevTool Detail<br/>
-개발 툴 세부내용 페이지입니다. 자신이 다룰 수 있는 개발툴에 대한 상세목록을 확인해보세요.
-<br/>
-<h1>{{ tool.name }}</h1>
-<br/>
-    {{ tool.kind }}
-    <br/>
-    {{ tool.description }}
-    <br/>
-    {%for idea in tool.idea_set.all %}
-        <a href="{% url 'ideas:idea_detail' idea.pk%}">{{ idea.title }}</a>
-    {% endfor %}
-
-    <button class="btn" onclick="location.href='{% url 'devtoolBlog:delete_devtool' tool.pk %}'">삭제</button>
-    <button class="btn" onclick="location.href='{% url 'devtoolBlog:edit_devtool' tool.pk %}'">수정</button>
-    <button class="btn" onclick="location.href='{% url "devtoolBlog:devtool_list" %}'">목록</button>
