--- SWIDEA_SITE/devtoolBlog/templates/list.html
@@ -1,8 +1,0 @@
-<a href="">DevTool List</a><br/>
-개발 툴 목록 페이지입니다. 자신이 사용가능한 개발툴의 목록을 확인해보세요.
-<br/>
-{%for devtool in devtool %}
-    <a href="{% url 'devtoolBlog:devtool_detail' tool.pk%}"><h1>{{devtool.name}}</h1></a>
-    <p>{{ devtool.kind }}</p>
-    <p>{{ devtool.description }}</p>
-{% endfor %}
