+++ SWIDEA_SITE/ideas/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class IdeablogConfig(AppConfig):
+class IdeasConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'ideaBlog'
+    name = 'ideas'
